Django/requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.11.0
django-filter==24.1
django-qrcode==0.3
django-rest-swagger==2.2.0
django-simpleui==2023.12.12
djangorestframework==3.15.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.2.2
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
inflection==0.5.1
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==3.0.2
mysql==0.0.3
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
notebook==7.1.2
notebook_shim==0.2.4
openapi-codec==1.3.2
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
Pillow==9.4.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.3.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse @ file:///C:/b/abs_88361ub_qu/croot/sqlparse_1690904577514/work
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.13
utils==1.0.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.2
widgetsnbextension==4.0.10
